Fix for the travel app
