+++ travel-app/App.js
@@ -1,5 +1,4 @@
-import React from 'react';
 import { createAppContainer } from 'react-navigation';
-import Travel from './navigation/Travel';
+import Travel from './src/navigation/Travel';
 
 export default createAppContainer(Travel);
