+++ travel-app/mocks.json
@@ -1,0 +1,81 @@
+[
+    {
+        "id": 1,
+        "user": {
+            "name": "Lelia Chavez",
+            "avatar": "https://randomuser.me/api/portraits/women/44.jpg"
+        },
+        "saved": true,
+        "location": "Santorini, Greece",
+        "temperature": 34,
+        "title": "Santorini",
+        "description": "Santorini is one of the Cyclades islands in the Aegean Sea. It was devastated by a volcanic eruption in the 16th century BC, forever shaping its rugged landscape. The whitewashed, cubiform houses of its 2 principal towns, Fira and Oia, cling to cliffs above an underwater caldera (crater). They overlook the sea, small islands to the west and beaches made up of black, red and white lava pebbles.",
+        "rating": 4.3,
+        "reviews": 3212,
+        "preview": "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80",
+        "images": [
+            "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80",
+            "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80",
+            "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80",
+            "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80"
+        ]
+    },
+    {
+        "id": 2,
+        "user": {
+            "name": "Lelia Chavez",
+            "avatar": "https://randomuser.me/api/portraits/women/44.jpg"
+        },
+        "saved": false,
+        "location": "Loutraki, Greece",
+        "temperature": 34,
+        "title": "Loutraki",
+        "description": "This attractive small town, 80 kilometers from Athens",
+        "rating": 4.6,
+        "reviews": 3212,
+        "preview": "https://images.unsplash.com/photo-1458906931852-47d88574a008?auto=format&fit=crop&w=800&q=80",
+        "images": [
+            "https://images.unsplash.com/photo-1458906931852-47d88574a008?auto=format&fit=crop&w=800&q=80",
+            "https://images.unsplash.com/photo-1446903572544-8888a0e60687?auto=format&fit=crop&w=800&q=80"
+        ]
+    },
+    {
+        "id": 3,
+        "user": {
+            "name": "Lelia Chavez",
+            "avatar": "https://randomuser.me/api/portraits/women/44.jpg"
+        },
+        "saved": true,
+        "location": "Santorini, Greece",
+        "temperature": 34,
+        "title": "Santorini",
+        "description": "Santorini - Description",
+        "rating": 3.2,
+        "reviews": 3212,
+        "preview": "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80",
+        "images": [
+            "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80",
+            "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80",
+            "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80",
+            "https://images.unsplash.com/photo-1507501336603-6e31db2be093?auto=format&fit=crop&w=800&q=80"
+        ]
+    },
+    {
+        "id": 4,
+        "user": {
+            "name": "Lelia Chavez",
+            "avatar": "https://randomuser.me/api/portraits/women/44.jpg"
+        },
+        "location": "Loutraki, Greece",
+        "temperature": 34,
+        "title": "Loutraki",
+        "description": "This attractive small town, 80 kilometers from Athens",
+        "rating": 5,
+        "reviews": 3212,
+        "preview": "https://images.unsplash.com/photo-1458906931852-47d88574a008?auto=format&fit=crop&w=800&q=80",
+        "images": [
+            "https://images.unsplash.com/photo-1458906931852-47d88574a008?auto=format&fit=crop&w=800&q=80",
+            "https://images.unsplash.com/photo-1446903572544-8888a0e60687?auto=format&fit=crop&w=800&q=80"
+        ]
+    }
+]
