+++ travel-app/src/Theme.js
@@ -1,3 +1,4 @@
+
 const colors = {
   black: '#000',
   white: '#FFF',
@@ -12,8 +13,7 @@ const sizes = {
   padding: 36,
   margin: 36,
   title: 24,
-  border: 16,
-  radius: 12,
+  radius: 8,
 };
 
 export {
