+++ travel-app/src/navigation/Travel.js
@@ -1,8 +1,7 @@
-import React from 'react';
 import { createStackNavigator } from 'react-navigation';
 
-import List from '../screens/List';
-import Article from '../screens/Article';
+import List from '../screens/list/List';
+import Article from '../screens/articles/Article';
 
 export default createStackNavigator(
   {
