+++ travel-app/src/screens/articles/Article.js
@@ -1,91 +1,19 @@
-import React, { Component } from 'react'
-import { Text, StyleSheet, View, Animated, Image, Dimensions, ScrollView, TouchableOpacity } from 'react-native'
-
+import React, { Component } from 'react';
+import { 
+    Text,
+    View,
+    Animated,
+    Image,
+    Dimensions,
+    ScrollView,
+    TouchableOpacity
+} from 'react-native';
 import FontAwesome from 'react-native-vector-icons/FontAwesome';
 import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
+import * as theme from '../../Theme';
+import styles from './Article.styles';
 
-import * as theme from '../theme';
-
-const { width, height } = Dimensions.get('window');
-
-const styles = StyleSheet.create({
-  flex: {
-    flex: 1,
-  },
-  column: {
-    flexDirection: 'column'
-  },
-  row: {
-    flexDirection: 'row'
-  },
-  header: {
-    // backgroundColor: 'transparent',
-    paddingHorizontal: theme.sizes.padding,
-    paddingTop: theme.sizes.padding,
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  back: {
-    width: theme.sizes.base * 3,
-    height: theme.sizes.base * 3,
-    justifyContent: 'center',
-    alignItems: 'flex-start',
-  },
-  content: {
-    // backgroundColor: theme.colors.active,
-    // borderTopLeftRadius: theme.sizes.border,
-    // borderTopRightRadius: theme.sizes.border,
-  },
-  contentHeader: {
-    backgroundColor: 'transparent',
-    padding: theme.sizes.padding,
-    backgroundColor: theme.colors.white,
-    borderTopLeftRadius: theme.sizes.border,
-    borderTopRightRadius: theme.sizes.border,
-    marginTop: -theme.sizes.padding / 2,
-  },
-  avatar: {
-    position: 'absolute',
-    top: -theme.sizes.margin,
-    right: theme.sizes.margin,
-    width: theme.sizes.padding * 2,
-    height: theme.sizes.padding * 2,
-    borderRadius: theme.sizes.padding,
-  },
-  shadow: {
-    shadowColor: theme.colors.black,
-    shadowOffset: {
-      width: 0,
-      height: 6,
-    },
-    shadowOpacity: 0.5,
-    shadowRadius: 5,
-  },
-  dotsContainer: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    position: 'absolute',
-    bottom: 36,
-    right: 0,
-    left: 0
-  },
-  dots: {
-    width: 8,
-    height: 8,
-    borderRadius: 4,
-    marginHorizontal: 6,
-    backgroundColor: theme.colors.gray,
-  },
-  title: {
-    fontSize: theme.sizes.font * 2,
-    fontWeight: 'bold'
-  },
-  description: {
-    fontSize: theme.sizes.font * 1.2,
-    lineHeight: theme.sizes.font * 2,
-    color: theme.colors.caption
-  }
-});
+const { width } = Dimensions.get('window');
 
 class Article extends Component {
   scrollX = new Animated.Value(0);
